Creation of hallway building blocks
Sean began to make some basic components for building hallways. Currently, a regular hallway and a corner have been created. To combine them, the use of math based on their measurements is needed with their positions. Based on the measurements, a measure of about 12.5 will be the most common measurement for combining the pieces.